Create core entities for task list (#2)
Add the core entity to the project, that is the Task itself. It also
includes the logic of some basic business operations on the entity and
some of the parameters that will be included in the ML model training.

Also, add a base model for all entities and unit tests.